requirements.txt
django==3.0.5
psycopg2==2.8.5
django-taggit
Faker==4.0.3
python-dotenv
dj-database-url
uwsgi==2.0.18
werkzeug==1.0.1
whitenoise==5.0.1
Pillow==7.1.2
django-heroku==0.3.1
requests
django-mptt
django-tinymce-4
newspaper3k